doc, tools: extrat buildToc
