+++ tools/doc/html.js
@@ -18,6 +18,7 @@ const parseYAML = require('./lib/parseYAML')
 const linkJsTypeDocs = require('./lib/linkJsTypeDocs')
 const parseAPIHeader = require('./lib/parseAPIHeader')
 const getSection = require('./lib/getSection')
+const buildToc = require('./lib/buildToc')
 
 module.exports = toHTML;
 
@@ -30,31 +31,6 @@ marked.setOptions({
   renderer: renderer
 });
 
-
-
-
-function buildToc(lexed, filename, cb) {
-  var toc = [];
-  var depth = 0;
-  lexed.forEach(function(tok) {
-    if (tok.type !== 'heading') return;
-    if (tok.depth - depth > 1) {
-      return cb(new Error('Inappropriate heading level\n' +
-                          JSON.stringify(tok)));
-    }
-
-    depth = tok.depth;
-    var id = getId(filename + '_' + tok.text.trim());
-    toc.push(new Array((depth - 1) * 2 + 1).join(' ') +
-             '* <a href="#' + id + '">' +
-             tok.text + '</a>');
-    tok.text += '<span><a class="mark" href="#' + id + '" ' +
-                'id="' + id + '">#</a></span>';
-  });
-
-  toc = marked.parse(toc.join('\n'));
-  cb(null, toc);
-}
 
 var idCounters = {};
 function getId(text) {
