+++ tools/doc/lib/buildToc.js
@@ -1,0 +1,23 @@
+
+module.exports = function buildToc(lexed, filename, cb) {
+  var toc = [];
+  var depth = 0;
+  lexed.forEach(function(tok) {
+    if (tok.type !== 'heading') return;
+    if (tok.depth - depth > 1) {
+      return cb(new Error('Inappropriate heading level\n' +
+                          JSON.stringify(tok)));
+    }
+
+    depth = tok.depth;
+    var id = getId(filename + '_' + tok.text.trim());
+    toc.push(new Array((depth - 1) * 2 + 1).join(' ') +
+             '* <a href="#' + id + '">' +
+             tok.text + '</a>');
+    tok.text += '<span><a class="mark" href="#' + id + '" ' +
+                'id="' + id + '">#</a></span>';
+  });
+
+  toc = marked.parse(toc.join('\n'));
+  cb(null, toc);
+}
